requirements.txt
flask
requests
beautifulsoup4
pysnmp
paramiko
python-telegram-bot
flask_sqlalchemy
Flask-Migrate
